Create working load.py file draft
Created load.py which takes most recent csv in
the "out" folder and uploads it to BigQuery as
a table. It is set to ifexsists='replace' for
now.

Next, we must refine the Transform to be more
useful for data analysis (change months to
numbers and create the coverage table). Then,
upgrade the Load function to only append Primary
Keys that do not exist in the table yet (may
involve reading from BigQuery).